+++ app/src/main/java/com/tolstykh/eatABurrita/data/BurritoEntry.kt
@@ -1,9 +1,10 @@
 package com.tolstykh.eatABurrita.data
 
 import androidx.room.Entity
+import androidx.room.Index
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "burrito_entries")
+@Entity(tableName = "burrito_entries", indices = [Index(value = ["timestamp"])])
 data class BurritoEntry(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val timestamp: Long,